refactor: split IslandView.swift into per-concern view files
Stage 2 of #47. Mechanical move only — every view definition kept
byte-identical, just relocated under `Sources/DynamicIsland/Views/`
with focused per-concern files. No behaviour change.

Before: IslandView.swift ~1183 lines housing 18 different SwiftUI
view types from Root → ears → expanded → capsule → pulse → detail
→ reply → progress, in one scrolling slab.

After:

- `IslandView.swift` (117 lines) — `IslandRootView` only
- `Views/Ears.swift` (233) — Left/Right ear shapes + ear views
- `Views/ExpandedContent.swift` (99) — notch's expanded panel
- `Views/Capsule.swift` (214) — fallback compact + expanded pills
- `Views/Pulse.swift` (115) — thinking pulse + thinking pill
- `Views/Detail.swift` (142) — diff detail + session tree + row
- `Views/Reply.swift` (229) — pending dots, permission buttons,
  quick replies, inline reply field
- `Views/Progress.swift` (41) — linear bar + ring

Per-file size now bounded under 250 lines each; future feature
work touches a clearly-named file instead of scrolling the
1100-line monolith.

`swift test`: 217 tests, 0 failures (unchanged from Stage 1 — this
commit is pure code reorg).
